+++ src/packages/@ncigdc/modern_components/ClinicalAnalysis/ClinicalVariableCard/index.js
@@ -1,4 +1,4 @@
-import CategoricalVariableCard from './EnhancedCategorical';
-import ContinuousVariableCard from './EnhancedContinuous';
+import CategoricalVariableCard from './CategoricalVariableCard';
+import ContinuousVariableCard from './ContinuousVariableCard';
 
 export { CategoricalVariableCard, ContinuousVariableCard };
