+++ src/packages/@ncigdc/modern_components/ClinicalAnalysis/ClinicalVariableCard/modals/CategoricalCustomBinsModal/index.js
@@ -1,0 +1,3 @@
+import CategoricalCustomBinsModal from './CategoricalCustomBinsModal';
+
+export default CategoricalCustomBinsModal;
